+++ src/server/game/Achievements/AchievementMgr.cpp
@@ -2307,7 +2307,7 @@ void AchievementMgr<T>::CompletedAchievement(AchievementEntry const* achievement
             draft.AddItem(item);
         }
 
-        draft.SendMailTo(trans, GetOwner(), MailSender(MAIL_CREATURE, reward->sender));
+        draft.SendMailTo(trans, GetOwner(), MailSender(MAIL_CREATURE, uint64(reward->sender)));
         CharacterDatabase.CommitTransaction(trans);
     }
